requirements.txt
Django==3.1.3
pytz==2020.4
sqlparse==0.4.1
gunicorn
djangorestframework==3.11.1
python-decouple
psycopg2-binary
python-memcached
celery>=4.4, <4.5
requests
beautifulsoup4==4.9.1
Pillow